Add files via upload
his research delves into the phenomena of breathers and wave-trains within the
context of internal gravity waves. The study begins with an introduction to internal
gravity waves and their importance in oceanography. Subsequently, attention is directed
towards the Gardner and modified Korteweg-de Vries (mKdV) equations, serving as the
governing mathematical equations used to model a special type of internal wave called
’breathers’. Additionally, internal breathers and wave-trains are presented, and their
distinguishing features are examined.