Add files via upload
In this report we explore the effects of a rotating frame of reference on the dynamics
of fluids. First we introduce a general rotation on basis vectors in the inertial reference
frame to produce the ficticious forces that arise in a non-inertrial reference frame. A
simpler rotating frame of reference is then examined to gain insight on the Coriolis and
centrifugal ficticious forces. A dimensional analysis argument is then created to explore
the regimes in which the Coriolis force and viscosity force are negligible. Next we focus
on a special case of geostrophic flow which is analyzed and a derivation of the Taylor-
Proudman Theorem is utilized to show the existence of Taylor Columns. We then attempt
to show how geostrophic flow works in the atmosphere where density is not constant .
Finally we reintroduce viscosity to describe the flow near a boundary in a rotating frame.